inline XFormService implementation
